+++ src/main/kotlin/com/jfrog/conan/clionplugin/toolWindow/ConanWindowFactory.kt
@@ -23,7 +23,6 @@ import com.intellij.ui.components.JBLabel
 import com.intellij.ui.components.JBPanelWithEmptyText
 import com.intellij.ui.components.JBScrollPane
 import com.intellij.ui.content.ContentFactory
-import com.intellij.ui.jcef.JCEFHtmlPanel
 import com.intellij.ui.table.JBTable
 import com.intellij.util.text.SemVer
 import com.intellij.util.ui.JBUI
@@ -69,10 +68,7 @@ class ConanWindowFactory : ToolWindowFactory {
 
             secondComponentPanel.layout = BoxLayout(secondComponentPanel, BoxLayout.Y_AXIS)
 
-            val htmlPanel = JCEFHtmlPanel(null).apply {
-                alignmentX = Component.LEFT_ALIGNMENT
-                loadHTML("")
-            }
+            val packageInfo = PackageInfoPanel()
 
             firstComponent = DialogPanel(BorderLayout()).apply {
                 border = JBUI.Borders.empty(5)
@@ -253,59 +249,11 @@ class ConanWindowFactory : ToolWindowFactory {
                         }
                         secondComponentPanel.add(buttonsPanel)
 
-                        val resourceFile = ConanWindowFactory::class.java.classLoader.getResource("conan/targets-data.json")
-                        val targetsData = resourceFile?.readText() ?: "{}"
 
-                        val script = """
-                            function fillExtraData() {
-                                const data = $targetsData
-                                const cpp_info = document.getElementById("cpp_info");
-                                cpp_info.innerText = ""
-                                if ("$name" in data) {
-                                    if ("cmake_file_name" in data["$name"]) {
-                                        cpp_info.innerText += "cmake_file_name: " + data["$name"]["cmake_file_name"] + "\n";
-                                    }
-                                    if ("cmake_target_name" in data["$name"]) {
-                                        cpp_info.innerText += "cmake_target_name: " + data["$name"]["cmake_target_name"] + "\n";
-                                    }
-                                    if ("components" in data["$name"]) {
-                                        cpp_info.innerText += "components: ";
-                                        for (let component in data["$name"]["components"]) {
-                                            cpp_info.innerText += data["$name"]["components"][component]["cmake_target_name"] + "\n";
-                                        }
-                                    }
-                                }
-                                else {
-                                    cpp_info.innerText += "cmake_file_name: $name\n";
-                                    cpp_info.innerText += "cmake_target_name: $name::$name\n";
-                                }
-                            }
-                        """.trimIndent()
-
-                        val htmlContent = """
-                            <html>
-                            <head>
-                                <style>
-                                    body {
-                                        color: rgb(187, 187, 187);
-                                        font-family: sans-serif;
-                                    }
-                                </style>
-                                <script>
-                                    $script
-                                </script>
-                            </head>
-                            <body onload="fillExtraData()">
-                                <div id="cpp_info"></div>
-                            </body>
-                            </html>
-                        """.trimIndent()
-
-                        htmlPanel.loadHTML(htmlContent)
 
                         secondComponentPanel.add(JPanel(FlowLayout(FlowLayout.LEFT)).apply {
                             alignmentX = Component.LEFT_ALIGNMENT
-                            add(htmlPanel.component)
+                            add(packageInfo.getComponent(name))
                         })
 
                         secondComponentPanel.revalidate()
